Update for mc 1.21.11, bump version to 1.6.0-rc1

+++ core/src/main/java/io/github/lumine1909/customworldheight/config/Height.java
@@ -3,6 +3,6 @@ package io.github.lumine1909.customworldheight.config;
 import java.util.Optional;
 import java.util.function.Function;
 
-public record Height(int height, int minY, int logicalHeight, Function<Optional<Integer>, Optional<Integer>> couldHeightFunc) {
+public record Height(int height, int minY, int logicalHeight, Function<Optional<Float>, Optional<Float>> couldHeightFunc) {
 
 }
